Some code cleanup.
One file per class.
Removed unnecessary exposed methods
