+++ src/Sidekick.Business/Apis/PoeNinja/IPoeNinjaCache.cs
@@ -1,17 +1,9 @@
-using Microsoft.Extensions.Logging;
-using Sidekick.Business.Trades.Results;
-using Sidekick.Core.Initialization;
-using Sidekick.Business.Apis.PoeNinja.Models;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using Sidekick.Business.Apis.PoeNinja.Models;
 using Sidekick.Business.Parsers.Models;
 
 namespace Sidekick.Business.Apis.PoeNinja
 {
-
     /// <summary>
     /// poe.ninja cache. The basic idea is fetching current poe.ninja with specified interval (e.g. hourly) in the background.
     /// imo it'd be overkill to request their api every time. Also perfomance. 
@@ -21,9 +13,6 @@ namespace Sidekick.Business.Apis.PoeNinja
     {
         DateTime? LastRefreshTimestamp { get; }
 
-        Task Refresh();
-
         PoeNinjaItem GetItem(Item item);
-
     }
 }
