+++ src/Sidekick.Business/Apis/PoeNinja/Models/Modifier.cs
@@ -1,0 +1,9 @@
+namespace Sidekick.Business.Apis.PoeNinja.Models
+{
+    public class Modifier
+    {
+        public string Text { get; set; }
+
+        public bool Optional { get; set; }
+    }
+}
