+++ src/Sidekick.Business/Apis/PoeNinja/Models/PoeNinjaCurrency.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Text.Json.Serialization;
-
 namespace Sidekick.Business.Apis.PoeNinja.Models
 {
     public class PoeNinjaCurrency : PoeNinjaResult
@@ -20,33 +17,5 @@ namespace Sidekick.Business.Apis.PoeNinja.Models
         public SparkLine LowConfidenceReviceSparkLine { get; set; }
 
         public string DetailsId { get; set; }
-    }
-
-    public class PoeNinjaExchange
-    {
-        public int Id { get; set; }
-
-        [JsonPropertyName("league_id")]
-        public int LeagueId { get; set; }
-
-        [JsonPropertyName("pay_currency_id")]
-        public int PayCurrencyId { get; set; }
-
-        [JsonPropertyName("get_currency_id")]
-        public int GetCurrencyId { get; set; }
-
-        [JsonPropertyName("sample_time_utc")]
-        public DateTime SampleTimeUtc { get; set; }
-
-        public int Count { get; set; }
-
-        public double Value { get; set; }
-
-        [JsonPropertyName("data_point_count")]
-        public int DataPointCount { get; set; }
-
-        [JsonPropertyName("includes_secondary")]
-        public bool IncludesSecondary { get; set; }
-
     }
 }
